changed `typealias Percentage = CGFloat` to `typealias Percentage = Float`. Added instance property `cgFloat` to `Percentage` that converts `self` to a `CGFloat`. `cgFloat` property is syntactically analogous to `cgColor` instance property of `UIColor`

+++ HLTChallenge/TypeAliases.swift
@@ -22,4 +22,4 @@ typealias ResultBlock<T> = (Result<T>) -> Void
 
 // MARK: - Percentage
 
-typealias Percentage = CGFloat
+typealias Percentage = Float
